requirements.txt
streamlit
plotly
requests
fastapi
pydantic
python-dotenv

